city, state and country master

+++ src/main/java/com/maan/eway/bean/CityMasterId.java
@@ -30,7 +30,7 @@ public class CityMasterId implements Serializable {
     private static final long serialVersionUID = 1L;
 
     //--- ENTITY KEY ATTRIBUTES 
-	private Long cityId;
+	private Integer cityId;
 
 	private Date effectiveDateStart;
